# docs/openapi.yaml
openapi: 3.0.3
info:
  title: F4ALL Sports Talent Assessment API
  description: >
    API contract for the F4ALL AI-powered sports talent assessment platform.
    The mobile application receives a provisional on-device score, while the
    backend performs server-side verification and stores the final result.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /api/athletes/register:
    post:
      tags: [Athletes]
      summary: Create a profile for a verified phone
      description: >
        Requires the registering token returned by verify-otp when registered
        is false. The phone is taken from that token, never from the body.
        Returns a fresh athlete session; the registering one is revoked.
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, dob, gender, region]
              properties:
                name: {type: string, maxLength: 150}
                dob: {type: string, format: date}
                gender: {type: string, enum: [male, female, other]}
                region:
                  type: string
                  description: An Indian state or union territory, canonical spelling
                height_cm: {type: number, nullable: true}
                weight_kg: {type: number, nullable: true}
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  profile: {$ref: "#/components/schemas/AthleteProfile"}
                  tokens: {$ref: "#/components/schemas/TokenPair"}
        "403": {description: Token is not a registering token}
        "409": {description: Phone already registered}
        "422": {description: Invalid age, gender or region}
  /api/athletes/me:
    get:
      tags: [Athletes]
      summary: The calling athlete's profile
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AthleteProfile"}
    patch:
      tags: [Athletes]
      summary: Update name, region, height or weight
      description: Date of birth and gender select the benchmark cohort and are not self-service.
      security: [{bearerAuth: []}]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: {type: string}
                region: {type: string}
                height_cm: {type: number}
                weight_kg: {type: number}
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AthleteProfile"}
  /api/athletes/me/photo:
    post:
      tags: [Athletes]
      summary: Upload the registration photo used by the identity check
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: {type: string, format: binary}
      responses:
        "200": {description: Saved; any previous photo is deleted}
        "413": {description: Larger than 5MB}
        "415": {description: Not JPEG, PNG or WebP}
  /api/athletes/me/summary:
    get:
      tags: [Athletes]
      summary: Profile, full result history and personal bests
      description: Personal bests count only verified or approved results, and say which are official.
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Summary}

  # -------------------------------------------------------------------------
  # Sprint 8: official dashboard
  # -------------------------------------------------------------------------
...
components:
...
  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT


  # ------------------------------------------------------------
  # REQUEST SCHEMAS
  # ------------------------------------------------------------
  schemas:
...
    TokenPair:
      type: object
      properties:
        access_token: {type: string}
        refresh_token: {type: string}
        token_type: {type: string, example: bearer}
        expires_in: {type: integer, description: Access token lifetime in seconds}
        athlete_id: {type: string, format: uuid, nullable: true}
        registered: {type: boolean}
    AthleteProfile:
      type: object
      properties:
        athlete_id: {type: string, format: uuid}
        name: {type: string}
        dob: {type: string, format: date}
        age_years: {type: integer}
        gender: {type: string}
        region: {type: string}
        phone: {type: string}
        height_cm: {type: number, nullable: true}
        weight_kg: {type: number, nullable: true}
        has_reference_photo: {type: boolean}